Validate sample fields plugin replaced onSave plugin.
